requirements.txt
Django==5.0.1
python-dotenv==1.0.1
python-dateutil==2.8.2
